Fix awk's complaints about too many fields
I switched some of the awk calls with simple calls to "cut", which
allowed me to narrow th error down to _new_only(), at which point
I realized that awk was trying to process fields delimited by ANY
WHITESPACE, which would lead to thousands of different fields--exactly
what it was complaining about.  I've set the field separator to '\t',
and all the errors have vanished into the aether.